refactored to remove duplicate functions and modularized
