+++ pages/history.jsx
@@ -6,7 +6,7 @@ import { AppBar, Button, Tab, Tabs, Grid, Typography, Card, Select, MenuItem, Li
 import axios from 'axios';
 import { withStyles } from '@material-ui/core/styles';
 import { lines, linesByName } from '../helpers/LineInfo.js';
-import { prepareNetworkData, dateToString } from "../helpers/PrepareData"
+import { dateToString, deriveLine, deriveXAxis, deriveYAxis, prepareTableData } from "../helpers/formatHistory"
 import PropTypes from 'prop-types';
 
 const styles = theme => ({
@@ -54,78 +54,13 @@ const styles = theme => ({
   }
 });
 
-const deriveLine = (data, line) => {
-    if (line === "All Lines") {
-      return data.map(datum => prepareNetworkData(datum))
-    } else {
-      const lineNum = linesByName[line].id;
-      return data.map(datum => datum[`${lineNum}_lametro-rail`]);
-    }
-}
-
-const deriveXAxis = (data, axisLabel) => {
-  switch(axisLabel) {
-    case "Weekday Average":
-      let weekDayArr = [[],[],[],[],[],[],[]]
-      for (let item of data) {
-        weekDayArr[new Date(item.date).getDay()].push(item);
-      }
-      const formattedWeekDayArr = weekDayArr.map(item => getAverageStats(item));
-      return formattedWeekDayArr;
-    case "Last 30 Days":
-      const lastThirtyArr = data.slice(data.length-30, data.length);
-      const formattedLastThirtyArr = lastThirtyArr.map(currItem => ({
-        wait: currItem.mean_time_between / 60,
-        oneMin: currItem.ontime["1_min"] / currItem.total_arrivals_analyzed * 100,
-        fiveMin: currItem.ontime["5_min"] / currItem.total_arrivals_analyzed * 100,
-      }))
-      return formattedLastThirtyArr;
-    default:
-       return data;
-  }
-}
-
-const getAverageStats = (arr) => {
-  const wait = arr.reduce((acc, currItem) => acc + currItem.mean_time_between, 0) / arr.length / 60;
-  const oneMin = arr.reduce((acc, currItem) => acc + currItem.ontime["1_min"] / currItem.total_arrivals_analyzed, 0) / arr.length * 100;
-  const fiveMin = arr.reduce((acc, currItem) => acc + currItem.ontime["5_min"] / currItem.total_arrivals_analyzed, 0) / arr.length * 100;
-  return {
-    wait,
-    oneMin,
-    fiveMin
-  }
-}
-
-const deriveYAxis = (data, axisLabel) => {
-  switch(axisLabel) {
-    case "Average Wait Time":
-      return data.map((item, i) => {
-        return item.wait;
-      });
-    case "% Within 1 Minute":
-      return data.map((item, i) => {
-        return item.oneMin;        
-      });
-    case "% Within 5 Minutes":
-    return data.map((item, i) => {
-        return item.fiveMin;
-      });
-    default:
-       return data;
-  }
-}
-
-const prepareHistoryData = (data) => {
-    return data.slice(0,data.length).reverse().map((item,i) => Object.assign(item,{ id: i }))
-}
-
 class History extends React.Component {
   state = {
     rows: [],
     graphData: [],
     line: "All Lines",
-    xAxis: "Weekday Average",
-    xTickFormat: ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"],
+    xAxis: "Last 30 Days",
+    xTickFormat: new Array(30).fill("").map((item, i) => dateToString(30-i)),
     yAxis: "Average Wait Time",
     yTickFormat: { formatter: function() { return `${this.value} min`; } },
     dataFormat: "chart",
@@ -135,44 +70,43 @@ class History extends React.Component {
 
   static async getInitialProps({ query, res }) {
     const { data } = await axios.get('http://localhost:8080/history');
-    const formattedData = Object.values(data);
-    return { query, formattedData };
+    const formattedData = Object.values(data[0]);
+    const allLineData = data[1];
+    return { query, allLineData, formattedData };
   }
 
   static getDerivedStateFromProps (props, state) {
     if (!state.rows[0] && !state.graphData[0]) {
       return {
-        rows: prepareHistoryData(props.formattedData.map(datum => prepareNetworkData(datum))),
-        graphData: deriveYAxis(deriveXAxis(deriveLine(props.formattedData, state.line), state.xAxis), state.yAxis)
+        rows: prepareTableData(props.allLineData),
+        graphData: deriveYAxis(deriveXAxis(deriveLine(props.formattedData, state.line, props.allLineData), state.xAxis), state.yAxis)
       }
     }
     return null;
   }
 
-
   handleLineChange = event => {
     this.setState({ line: event.target.value,
                     color: linesByName[event.target.value]["color"],
-                    rows: prepareHistoryData(deriveLine(this.props.formattedData, event.target.value)),
-                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, event.target.value), this.state.xAxis), this.state.yAxis)
+                    rows: prepareTableData(deriveLine(this.props.formattedData, event.target.value, this.props.allLineData)),
+                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, event.target.value, this.props.allLineData), this.state.xAxis), this.state.yAxis)
                  });
   };
 
   handleXAxisChange = event => {
     this.setState({ xAxis: event.target.value,
-                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, this.state.line), event.target.value), this.state.yAxis),
+                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, this.state.line, this.props.allLineData), event.target.value), this.state.yAxis),
                  });
     if (event.target.value === "Weekday Average") {
       this.setState({ xTickFormat: ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"] });
     } else {
       this.setState({ xTickFormat: new Array(30).fill("").map((item, i) => dateToString(30-i)) });
     }
-    console.log(this.state.xTickFormat);
   };
 
   handleYAxisChange = event => {
     this.setState({ yAxis: event.target.value,
-                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, this.state.line), this.state.xAxis), event.target.value),
+                    graphData: deriveYAxis(deriveXAxis(deriveLine(this.props.formattedData, this.state.line, this.props.allLineData), this.state.xAxis), event.target.value),
                     
                  });
     if (event.target.value === "Average Wait Time") {
@@ -266,9 +200,9 @@ class History extends React.Component {
             name="xAxis"
             className={classes.selectEmpty}
           >
-            <MenuItem value={"Weekday Average"}>Weekday Average
-            </MenuItem>
             <MenuItem value={"Last 30 Days"}>Last 30 Days
+            </MenuItem>
+            <MenuItem value={"Weekday Average"}>Weekday Average
             </MenuItem>
           </Select>
           </div>
